add income's methods

+++ src/main/java/ru/telegram/entity/IncomesEntity.java
@@ -43,7 +43,5 @@ public class IncomesEntity {
      * Amount
      */
     @Column(name = "amount")
-    Integer amount;
-
-
+    Double amount;
 }
